[fix] try

+++ backend/database/views_avatar.py
@@ -4,7 +4,6 @@ from django.views.decorators.http import require_http_methods
 from .models import User
 
 
-@require_http_methods(["POST"])
 def edit_avatar(request):
     username = request.POST.get('username')
     avatar = request.FILES.get('avatar')
@@ -14,12 +13,14 @@ def edit_avatar(request):
         res = {'code': 401, 'message': '用户不存在'}
         return JsonResponse(res)
 
-    user.avatar.delete()
+    if user.avatar:
+        user.avatar.delete()
     user.avatar = avatar
     user.save()
 
     res = {'code': 200, 'message': '头像修改成功'}
     return JsonResponse(res)
+
 
 
 @require_http_methods(["POST"])
@@ -30,5 +31,10 @@ def get_avatar(request):  # 获取用户头像
         res = {'code': 401, 'message': '用户不存在'}
         return JsonResponse(res)
 
-    res = {'code': 200, 'message': '头像获取成功', 'avatar': user.avatar.url}
+    if user.avatar:
+        res = {'code': 200, 'message': '头像获取成功', 'avatar': user.avatar.url}
+    else:
+        res = {'code': 200, 'message': '头像获取成功', 'avatar': 'default.png'}
+
     return JsonResponse(res)
+
